Fix delete command abstraction to be similar to other commands, to adopt better SLAP principle
