+++ src/main/java/seedu/jarvis/logic/parser/DeleteCommandParser.java
@@ -1,13 +1,14 @@
 package seedu.jarvis.logic.parser;
 
 import static seedu.jarvis.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.jarvis.logic.commands.delete.DeleteCommand.MESSAGE_DELETE_USAGE;
 import static seedu.jarvis.logic.parser.CliSyntax.DELETE_TASK;
 
 import java.util.logging.Logger;
 
-import seedu.jarvis.MainApp;
 import seedu.jarvis.commons.core.LogsCenter;
 import seedu.jarvis.logic.commands.delete.DeleteCommand;
+import seedu.jarvis.logic.commands.delete.DeleteTaskCommand;
 import seedu.jarvis.logic.parser.exceptions.ParseException;
 import seedu.jarvis.model.flag.Flag;
 
@@ -16,7 +17,7 @@ import seedu.jarvis.model.flag.Flag;
  */
 public class DeleteCommandParser implements Parser<DeleteCommand> {
 
-    private static final Logger logger = LogsCenter.getLogger(MainApp.class);
+    private static final Logger logger = LogsCenter.getLogger(DeleteCommandParser.class);
 
     /**
      * Parses the given {@code String} of arguments in the context of the DeleteCommand
@@ -27,7 +28,7 @@ public class DeleteCommandParser implements Parser<DeleteCommand> {
         String trimmedArgs = args.trim();
         if (trimmedArgs.isEmpty()) {
             throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_DELETE_USAGE));
+                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, MESSAGE_DELETE_USAGE));
         }
 
         // split the string trimmedArgs with regex of one or more whitespace characters.
@@ -38,7 +39,7 @@ public class DeleteCommandParser implements Parser<DeleteCommand> {
             commandFlag = ParserUtil.parseFlag(nameKeywords[0]);
         } catch (ParseException ex) {
             throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_DELETE_USAGE));
+                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, MESSAGE_DELETE_USAGE));
         }
 
         switch(commandFlag.getFlag()) {
@@ -46,26 +47,17 @@ public class DeleteCommandParser implements Parser<DeleteCommand> {
             logger.info("DeleteCommandParser attempts to parse user's delete Task input");
             try {
                 String taskId = TaskCommandParser.parseDeleteTask(nameKeywords);
-                return new DeleteCommand(taskId);
+                return new DeleteTaskCommand(taskId);
 
             } catch (ParseException pe) {
                 throw new ParseException(
-                        String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_DELETE_TASK_USAGE), pe);
+                        String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteTaskCommand.MESSAGE_DELETE_TASK_USAGE), pe);
             }
 
-        //Will not be able to delete student becasue the flag is not allowed in Flag class.
-        default:
-            /*
-            logger.info("DeleteCommandParser attempts to parse user's delete student input by default");
-            try {
-                Index index = ParserUtil.parseIndex(args);
-                return new DeleteCommand(index);
+            // Add other cases of delete command here.
 
-            } catch (ParseException pe) {
-                throw new ParseException(
-                        String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_DELETE_USAGE), pe);
-            }*/
-            throw new ParseException("This stage should not be reached!");
+        default:
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, MESSAGE_DELETE_USAGE));
         }
     }
